pyproject.toml
[project]
name = "7-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=3.0.0",
    "ipykernel>=7.2.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.11",
    "langchain-tavily>=0.2.18",
    "langgraph>=1.1.3",
    "langgraph-cli[inmem]>=0.4.24",
    "python-dotenv>=1.2.2",
    "typing-extensions>=4.15.0",
    "wikipedia>=1.4.0",
]
